django_project/tmpl/_$project_name$_/requirements/common.txt
django==_$django_version$_
ipython

dj-database-url==0.2.2
django-dotenv==1.2
django-getenv==1.1
